Исправление краша по получение токена
