+++ restaurant/src/main/java/ru/papino/restaurant/App.kt
@@ -1,45 +1,16 @@
 package ru.papino.restaurant
 
 import android.app.Application
-import kotlinx.coroutines.CoroutineScope
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
-import ru.papino.restaurant.core.user.di.UserDI
-import ru.papino.restaurant.core.user.encrypted.EncryptedToken
-import ru.papino.restaurant.data.repository.UserRepositoryImpl
-import ru.papino.restaurant.domain.repository.models.UserResponse
-import ru.papino.restaurant.domain.usecases.GetUserByTokenUseCase
+import android.content.Context
 
 class App : Application() {
-    private val getUserByTokenUseCase =
-        GetUserByTokenUseCase(userRepository = UserRepositoryImpl.getInstance())
-
     override fun onCreate() {
         super.onCreate()
-        authorizationByToken()
+        _context = applicationContext
     }
 
-    private fun authorizationByToken() {
-        val token = EncryptedToken.getToken()
-        token?.let {
-            CoroutineScope(Dispatchers.IO).launch {
-                runCatching {
-                    getUserByTokenUseCase(it)
-                }.onSuccess { response ->
-                    when (response) {
-                        is UserResponse.Success -> {
-                            UserDI.init(response.user)
-                            UserDI.initToken(response.token)
-                        }
-
-                        is UserResponse.Error -> {
-                            UserDI.clear()
-                        }
-                    }
-                }.onFailure {
-                    UserDI.clear()
-                }
-            }
-        }
+    companion object {
+        private var _context: Context? = null
+        fun getApplicationContext() = _context
     }
 }
