+++ restaurant/src/main/java/ru/papino/restaurant/RestaurantActivity.kt
@@ -9,10 +9,16 @@ import androidx.core.content.res.ResourcesCompat
 import androidx.fragment.app.Fragment
 import androidx.lifecycle.lifecycleScope
 import com.google.android.material.badge.BadgeDrawable
+import kotlinx.coroutines.CoroutineScope
+import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import ru.papino.restaurant.core.room.RoomDependencies
 import ru.papino.restaurant.core.user.di.UserDI
+import ru.papino.restaurant.core.user.encrypted.EncryptedToken
+import ru.papino.restaurant.data.repository.UserRepositoryImpl
 import ru.papino.restaurant.databinding.ActivityRestaurantBinding
+import ru.papino.restaurant.domain.repository.models.UserResponse
+import ru.papino.restaurant.domain.usecases.GetUserByTokenUseCase
 import ru.papino.restaurant.extensions.switchFragment
 import ru.papino.uikit.extensions.fullscreen
 
@@ -24,6 +30,7 @@ class RestaurantActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         RoomDependencies.init(applicationContext)
+        authorizationByToken()
         window.fullscreen()
         super.onCreate(savedInstanceState)
 
@@ -114,6 +121,32 @@ class RestaurantActivity : AppCompatActivity() {
                     ru.papino.uikit.R.color.backgroundMenuColor
                 )
             )
+        }
+    }
+
+    fun authorizationByToken() {
+        val getUserByTokenUseCase =
+            GetUserByTokenUseCase(userRepository = UserRepositoryImpl.getInstance())
+        val token = EncryptedToken.getToken()
+        token?.let {
+            CoroutineScope(Dispatchers.IO).launch {
+                runCatching {
+                    getUserByTokenUseCase(it)
+                }.onSuccess { response ->
+                    when (response) {
+                        is UserResponse.Success -> {
+                            UserDI.init(response.user)
+                            UserDI.initToken(response.token)
+                        }
+
+                        is UserResponse.Error -> {
+                            UserDI.clear()
+                        }
+                    }
+                }.onFailure {
+                    UserDI.clear()
+                }
+            }
         }
     }
